+++ src/HuffmanTree.cpp
@@ -16,7 +16,7 @@ HuffmanTree::HuffmanTree(array<uint64_t, CHAR_COUNT> frequencies)
     : frequencies(frequencies) {
     priority_queue<HuffmanTreeNode *, vector<HuffmanTreeNode *>, NodeComparator>
         pqueue;
-    for (int i = 0; i < this->frequencies.size(); i++) {
+    for (size_t i = 0; i < this->frequencies.size(); i++) {
         if (this->frequencies[i] > 0) {
             pqueue.push(new HuffmanTreeNode{static_cast<uint8_t>(i),
                                             this->frequencies[i], nullptr,
